MVP metrics visualizer
