+++ index.html
@@ -5,11 +5,7 @@
     <link
       rel="icon"
       type="image/svg+xml"
-      href="/kafka-sonar-transparent-logo.svg"
-    />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Montserrat&display=swap"
-      rel="stylesheet"
+      href="./public/kafka-sonar-transparent-logo.svg"
     />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Kafka Sonar</title>
