+++ src/App.css
@@ -1,3 +1,12 @@
+iframe {
+  height: 100%;
+  width: 100%;
+}
+
+#root {
+  font-family: 'Montserrat Variable', sans-serif;
+}
+
 /* #root {
   max-width: 1280px;
   margin: 0 auto;
