+++ src/App.tsx
@@ -1,35 +1,208 @@
-// import { useState } from 'react'
-// import reactLogo from './assets/react.svg'
-// import viteLogo from '/vite.svg'
-// import './App.css'
+// Baseline code reference: https://mui.com/material-ui/react-drawer/#mini-variant-drawer
+import * as React from 'react';
+import { styled, useTheme, Theme, CSSObject } from '@mui/material/styles';
+import { useNavigate } from 'react-router-dom';
+import Box from '@mui/material/Box';
+import MuiDrawer from '@mui/material/Drawer';
+import Toolbar from '@mui/material/Toolbar';
+import List from '@mui/material/List';
+import CssBaseline from '@mui/material/CssBaseline';
+import Divider from '@mui/material/Divider';
+import IconButton from '@mui/material/IconButton';
+import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
+import ChevronRightIcon from '@mui/icons-material/ChevronRight';
+import ListItem from '@mui/material/ListItem';
+import ListItemButton from '@mui/material/ListItemButton';
+import ListItemIcon from '@mui/material/ListItemIcon';
+import ListItemText from '@mui/material/ListItemText';
+import Typography from '@mui/material/Typography';
+// custom selected or created icons
+import HubIcon from '@mui/icons-material/Hub'; // cluster
+import DeviceHubIcon from '@mui/icons-material/DeviceHub'; // broker
+import PolylineIcon from '@mui/icons-material/Polyline'; // consumer
+import SvgIcon from '@mui/material/SvgIcon';
+import { ReactComponent as OrangeLogo } from './../src/assets/kafka-sonar-orange-logo.svg';
+// import { ReactComponent as WhiteLogo } from './../../public/kafka-sonar-white-logo.svg';
+// TS types
+import Props from './types/types';
+// Font supports weights 100-900
+import '@fontsource-variable/montserrat';
 
-// function App() {
-//   const [count, setCount] = useState(0)
+export default function App(props: Props) {
+  const drawerWidth = 200;
 
-//   return (
-//     <>
-//       <div>
-//         <a href="https://vitejs.dev" target="_blank">
-//           <img src={viteLogo} className="logo" alt="Vite logo" />
-//         </a>
-//         <a href="https://react.dev" target="_blank">
-//           <img src={reactLogo} className="logo react" alt="React logo" />
-//         </a>
-//       </div>
-//       <h1>Vite + React</h1>
-//       <div className="card">
-//         <button onClick={() => setCount((count) => count + 1)}>
-//           count is {count}
-//         </button>
-//         <p>
-//           Edit <code>src/App.tsx</code> and save to test HMR
-//         </p>
-//       </div>
-//       <p className="read-the-docs">
-//         Click on the Vite and React logos to learn more
-//       </p>
-//     </>
-//   )
-// }
+  const openedMixin = (theme: Theme): CSSObject => ({
+    backgroundColor: '#0d090a',
+    color: '#f8fbfd',
+    width: drawerWidth,
+    transition: theme.transitions.create('width', {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen,
+    }),
+    overflowX: 'hidden',
+  });
 
-// export default App
+  const closedMixin = (theme: Theme): CSSObject => ({
+    backgroundColor: '#0d090a',
+    color: '#f8fbfd',
+    transition: theme.transitions.create('width', {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen,
+    }),
+    overflowX: 'hidden',
+    width: `calc(${theme.spacing(7)} + 1px)`,
+    [theme.breakpoints.up('sm')]: {
+      width: `calc(${theme.spacing(8)} + 1px)`,
+    },
+  });
+
+  const Drawer = styled(MuiDrawer, {
+    shouldForwardProp: (prop) => prop !== 'open',
+  })(({ theme, open }) => ({
+    width: drawerWidth,
+    flexShrink: 0,
+    whiteSpace: 'nowrap',
+    boxSizing: 'border-box',
+    ...(open && {
+      ...openedMixin(theme),
+      '& .MuiDrawer-paper': openedMixin(theme),
+    }),
+    ...(!open && {
+      ...closedMixin(theme),
+      '& .MuiDrawer-paper': closedMixin(theme),
+    }),
+  }));
+
+  const {
+    brokerStats,
+    partitionStats,
+    brokerUtil,
+    brokerIO,
+    topicIO,
+    networkEff,
+  } = props;
+
+  const [open, setOpen] = React.useState(false);
+  const handleDrawerOpen = () => {
+    setOpen(true);
+  };
+  const handleDrawerClose = () => {
+    setOpen(false);
+  };
+
+  const navigate = useNavigate();
+
+  return (
+    <Box sx={{ display: 'flex' }}>
+      {/* <CssBaseline /> */}
+      <Drawer
+        variant="permanent"
+        open={open}
+        sx={{ backgroundColor: 'transparent' }}
+      >
+        <Toolbar>
+          <IconButton
+            color="inherit"
+            aria-label="open drawer"
+            onClick={handleDrawerOpen}
+            edge="start"
+            sx={{
+              marginRight: 5,
+              ...(open && { display: 'none' }),
+            }}
+          >
+            <ChevronRightIcon />
+          </IconButton>
+          <SvgIcon fontSize="large">
+            <OrangeLogo />
+          </SvgIcon>
+          {/* <SvgIcon fontSize="large" viewBox="0 0 1100 1100">
+            <WhiteLogo />
+          </SvgIcon> */}
+          <img
+            src="./../../public/kafka-sonar-white-logo.svg"
+            style={{ margin: 10, width: 75 }}
+          />
+          <IconButton
+            color="inherit"
+            aria-label="open drawer"
+            onClick={handleDrawerClose}
+            sx={{
+              ...(!open && { display: 'none' }),
+            }}
+          >
+            <ChevronLeftIcon />
+          </IconButton>
+        </Toolbar>
+        <Divider color="#f8fbfd" />
+        <List>
+          {['Cluster', 'Broker', 'Consumer'].map((text, index) => (
+            <ListItem
+              key={text}
+              disablePadding
+              onClick={
+                index === 0
+                  ? () => {
+                      navigate('/');
+                    }
+                  : index === 1
+                  ? () => {
+                      navigate('/broker');
+                    }
+                  : () => {
+                      navigate('/consumer');
+                    }
+              }
+              sx={{
+                display: 'block',
+              }}
+            >
+              <ListItemButton
+                sx={{
+                  minHeight: 48,
+                  justifyContent: open ? 'initial' : 'center',
+                  px: 2.5,
+                }}
+              >
+                <ListItemIcon
+                  sx={{
+                    minWidth: 0,
+                    mr: open ? 3 : 'auto',
+                    justifyContent: 'center',
+                    color: '#f8fbfd',
+                  }}
+                >
+                  {index === 0 ? (
+                    <HubIcon />
+                  ) : index === 1 ? (
+                    <DeviceHubIcon />
+                  ) : (
+                    <PolylineIcon />
+                  )}
+                </ListItemIcon>
+                <ListItemText
+                  primary={
+                    <Typography variant="inherit" fontFamily="inherit">
+                      {text}
+                    </Typography>
+                  }
+                  sx={{
+                    opacity: open ? 1 : 0,
+                  }}
+                />
+              </ListItemButton>
+            </ListItem>
+          ))}
+        </List>
+      </Drawer>
+      <Box component="main" sx={{ flexGrow: 1, p: 3 }}>
+        {partitionStats}
+        {brokerStats}
+        {networkEff}
+        {brokerIO}
+        {topicIO}
+        {brokerUtil}
+      </Box>
+    </Box>
+  );
+}
