--- src/components/NavDrawer.tsx
@@ -1,240 +1,0 @@
-// Baseline code reference: https://mui.com/material-ui/react-drawer/#mini-variant-drawer
-
-import * as React from 'react';
-import { styled, useTheme, Theme, CSSObject } from '@mui/material/styles';
-import { useNavigate } from 'react-router-dom';
-import Box from '@mui/material/Box';
-import MuiDrawer from '@mui/material/Drawer';
-// import MuiAppBar, { AppBarProps as MuiAppBarProps } from '@mui/material/AppBar';
-import Toolbar from '@mui/material/Toolbar';
-import List from '@mui/material/List';
-import CssBaseline from '@mui/material/CssBaseline';
-import Divider from '@mui/material/Divider';
-import IconButton from '@mui/material/IconButton';
-// import MenuIcon from '@mui/icons-material/Menu';
-import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
-import ChevronRightIcon from '@mui/icons-material/ChevronRight';
-import ListItem from '@mui/material/ListItem';
-import ListItemButton from '@mui/material/ListItemButton';
-import ListItemIcon from '@mui/material/ListItemIcon';
-import ListItemText from '@mui/material/ListItemText';
-
-import HubIcon from '@mui/icons-material/Hub'; // cluster
-import DeviceHubIcon from '@mui/icons-material/DeviceHub'; // broker
-import PolylineIcon from '@mui/icons-material/Polyline'; // consumer
-
-import SvgIcon from '@mui/material/SvgIcon';
-import { ReactComponent as OrangeLogo } from './../assets/kafka-sonar-orange-logo.svg';
-import BrokerStats from './BrokerStats';
-import PartitionStats from './PartitionStats';
-// import { ReactComponent as WhiteLogo } from './../../public/kafka-sonar-white-logo.svg';
-
-const drawerWidth = 200;
-
-const openedMixin = (theme: Theme): CSSObject => ({
-  backgroundColor: '#0d090a',
-  color: '#f8fbfd',
-  width: drawerWidth,
-  transition: theme.transitions.create('width', {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.enteringScreen,
-  }),
-  overflowX: 'hidden',
-});
-
-const closedMixin = (theme: Theme): CSSObject => ({
-  backgroundColor: '#0d090a',
-  color: '#f8fbfd',
-  transition: theme.transitions.create('width', {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  overflowX: 'hidden',
-  width: `calc(${theme.spacing(7)} + 1px)`,
-  [theme.breakpoints.up('sm')]: {
-    width: `calc(${theme.spacing(8)} + 1px)`,
-  },
-});
-
-// const DrawerHeader = styled('div')(({ theme }) => ({
-//   display: 'flex',
-//   alignItems: 'center',
-//   justifyContent: 'flex-end',
-//   padding: theme.spacing(0, 1),
-//   // necessary for content to be below app bar
-//   ...theme.mixins.toolbar,
-// }));
-
-// interface AppBarProps extends MuiAppBarProps {
-//   open?: boolean;
-// }
-
-// const AppBar = styled(MuiAppBar, {
-//   shouldForwardProp: (prop) => prop !== 'open',
-// })<AppBarProps>(({ theme, open }) => ({
-//   zIndex: theme.zIndex.drawer + 1,
-//   transition: theme.transitions.create(['width', 'margin'], {
-//     easing: theme.transitions.easing.sharp,
-//     duration: theme.transitions.duration.leavingScreen,
-//   }),
-//   ...(open && {
-//     marginLeft: drawerWidth,
-//     width: `calc(100% - ${drawerWidth}px)`,
-//     transition: theme.transitions.create(['width', 'margin'], {
-//       easing: theme.transitions.easing.sharp,
-//       duration: theme.transitions.duration.enteringScreen,
-//     }),
-//   }),
-// }));
-
-const Drawer = styled(MuiDrawer, {
-  shouldForwardProp: (prop) => prop !== 'open',
-})(({ theme, open }) => ({
-  width: drawerWidth,
-  flexShrink: 0,
-  whiteSpace: 'nowrap',
-  boxSizing: 'border-box',
-  ...(open && {
-    ...openedMixin(theme),
-    '& .MuiDrawer-paper': openedMixin(theme),
-  }),
-  ...(!open && {
-    ...closedMixin(theme),
-    '& .MuiDrawer-paper': closedMixin(theme),
-  }),
-}));
-
-interface Props {
-  brokerStats?: JSX.Element;
-  partitionStats?: JSX.Element;
-  brokerUtil?: JSX.Element;
-  brokerIO?: JSX.Element;
-  topicIO?: JSX.Element;
-}
-
-function NavDrawer(props: Props) {
-  const { brokerStats, partitionStats, brokerUtil, brokerIO, topicIO } = props;
-
-  const theme = useTheme();
-
-  const [open, setOpen] = React.useState(false);
-  const handleDrawerOpen = () => {
-    setOpen(true);
-  };
-  const handleDrawerClose = () => {
-    setOpen(false);
-  };
-
-  const navigate = useNavigate();
-
-  return (
-    <Box sx={{ display: 'flex' }}>
-      <CssBaseline />
-      <Drawer
-        variant="permanent"
-        open={open}
-        sx={{ backgroundColor: 'transparent' }}
-      >
-        <Toolbar>
-          <IconButton
-            color="inherit"
-            aria-label="open drawer"
-            onClick={handleDrawerOpen}
-            edge="start"
-            sx={{
-              marginRight: 5,
-              ...(open && { display: 'none' }),
-            }}
-          >
-            <ChevronRightIcon />
-          </IconButton>
-          <SvgIcon fontSize="large">
-            <OrangeLogo />
-          </SvgIcon>
-          {/* <SvgIcon fontSize="large" viewBox="0 0 1100 1100">
-            <WhiteLogo />
-          </SvgIcon> */}
-          <img
-            src="./../../public/kafka-sonar-white-logo.svg"
-            style={{ margin: 10, width: 75 }}
-          />
-          <IconButton
-            color="inherit"
-            aria-label="open drawer"
-            onClick={handleDrawerClose}
-            sx={{
-              ...(!open && { display: 'none' }),
-            }}
-          >
-            <ChevronLeftIcon />
-          </IconButton>
-        </Toolbar>
-        <Divider color="#f8fbfd" />
-        <List>
-          {['Cluster', 'Broker', 'Consumer'].map((text, index) => (
-            <ListItem
-              key={text}
-              disablePadding
-              onClick={
-                index === 0
-                  ? () => {
-                      navigate('/');
-                    }
-                  : index === 1
-                  ? () => {
-                      navigate('/broker');
-                    }
-                  : () => {
-                      navigate('/consumer');
-                    }
-              }
-              sx={{
-                display: 'block',
-              }}
-            >
-              <ListItemButton
-                sx={{
-                  minHeight: 48,
-                  justifyContent: open ? 'initial' : 'center',
-                  px: 2.5,
-                }}
-              >
-                <ListItemIcon
-                  sx={{
-                    minWidth: 0,
-                    mr: open ? 3 : 'auto',
-                    justifyContent: 'center',
-                    color: '#f8fbfd',
-                  }}
-                >
-                  {index === 0 ? (
-                    <HubIcon />
-                  ) : index === 1 ? (
-                    <DeviceHubIcon />
-                  ) : (
-                    <PolylineIcon />
-                  )}
-                </ListItemIcon>
-                <ListItemText
-                  primary={text}
-                  sx={{
-                    opacity: open ? 1 : 0,
-                  }}
-                />
-              </ListItemButton>
-            </ListItem>
-          ))}
-        </List>
-      </Drawer>
-      <Box component="main" sx={{ flexGrow: 1, p: 3 }}>
-        {brokerStats}
-        {brokerUtil}
-        {partitionStats}
-        {brokerIO}
-        {topicIO}
-      </Box>
-    </Box>
-  );
-}
-
-export default NavDrawer;
