+++ src/types/types.tsx
@@ -1,0 +1,8 @@
+export default interface Props {
+  brokerStats?: JSX.Element;
+  partitionStats?: JSX.Element;
+  brokerUtil?: JSX.Element;
+  brokerIO?: JSX.Element;
+  topicIO?: JSX.Element;
+  networkEff?: JSX.Element;
+}
